+++ bin/api_diff.ml
@@ -1,19 +1,12 @@
 let tool_name = "api-diff"
 
 type mode = Unwrapped | Wrapped of string | Cmi
-type word = Plain | Color
-type display = Line | Word of word
+type mark = [ `Plain | `Color ]
+type display = Line of mark | Word of mark
 
-let display_mode word =
-  match word with
-  | Some "plain" -> Ok (Word Plain)
-  | Some "color" -> Ok (Word Color)
-  | None -> Ok Line
-  | _ ->
-      Error
-        "Invalid argument. --word-diff takes an optional mode arguemnt.\n\
-         The mode can be either color or plain. If no mode is provided, it \
-         defaults to plain."
+let display_mode word_diff plain =
+  let mark = if plain then `Plain else `Color in
+  if word_diff then Word mark else Line mark
 
 let both_directories reference current =
   match (Sys.is_directory reference, Sys.is_directory current) with
@@ -55,13 +48,12 @@ let mode ~reference ~current ~main_module ~unwrapped =
 
 let print_diff text_diff display_mode =
   match display_mode with
-  | Line -> Api_watch.Text_diff.With_colors.pp Format.std_formatter text_diff
-  | Word Plain ->
-      Api_watch.Text_diff.Word.pp ~mode:`Plain Format.std_formatter text_diff
-  | Word Color ->
-      Api_watch.Text_diff.Word.pp ~mode:`Color Format.std_formatter text_diff
+  | Line mode ->
+      Api_watch.Text_diff.With_colors.pp ~mode Format.std_formatter text_diff
+  | Word mode ->
+      Api_watch.Text_diff.Word.pp ~mode Format.std_formatter text_diff
 
-let run (`Word_diff word_diff) (`Main_module main_module)
+let run (`Word_diff word_diff) (`Plain plain) (`Main_module main_module)
     (`Unwrapped_library unwrapped) (`Ref_cmi reference) (`Current_cmi current) =
   let open CCResult.Infix in
   let* reference_map, current_map =
@@ -93,7 +85,7 @@ let run (`Word_diff word_diff) (`Main_module main_module)
     |> List.filter_map (fun (_, v) -> v)
   in
   let has_changes = not (List.is_empty diff_map) in
-  let* display_mode = display_mode word_diff in
+  let display_mode = display_mode word_diff plain in
   List.iter
     (fun diff ->
       let text_diff = Api_watch.Text_diff.from_diff diff in
@@ -103,18 +95,19 @@ let run (`Word_diff word_diff) (`Main_module main_module)
 
 let named f = Cmdliner.Term.(app (const f))
 
-let word_diff =
-  let docv = "MODE" in
+let plain =
   let doc =
-    "Display the API diff in an inline word diff format rather than the usual \
-     line diff. Follows the same conventions as $(b,git diff --word-diff)."
+    "Add text markers to the output to highlight inline changes. Deleted parts \
+     are wrapped between [-and-] and added parts between {+and+}. Useful for \
+     terminals or outputs that don't support colors."
   in
+  named (fun x -> `Plain x) Cmdliner.Arg.(value & flag & info ~doc [ "plain" ])
+
+let word_diff =
+  let doc = "Display API changes in an inline word diff format." in
   named
     (fun x -> `Word_diff x)
-    Cmdliner.Arg.(
-      value
-      & opt ~vopt:(Some "plain") (some string) None
-      & info ~doc ~docv [ "word-diff" ])
+    Cmdliner.Arg.(value & flag & info ~doc [ "word-diff" ])
 
 let main_module =
   let docv = "MAIN_MODULE_NAME" in
@@ -163,7 +156,7 @@ let info =
 
 let term =
   Cmdliner.Term.(
-    const run $ word_diff $ main_module $ unwrapped_library $ ref_cmi
+    const run $ word_diff $ plain $ main_module $ unwrapped_library $ ref_cmi
     $ current_cmi)
 
 let () =