+++ tests/api-diff/word_based_diff_tests.t
@@ -24,37 +24,9 @@ We generate the .cmi file
 
   $ ocamlc cur.mli
 
-Run api-watcher on the two cmi file with default word-level diffing option enabled
+Run api-watcher on the two cmi file with word-level diffing flag enabled
 
-  $ api-diff ref.cmi cur.cmi --word-diff
-  diff module Cur:
-   type [-'a-] p = [-'a * 'a-]{+int * int+}
-   type t[- =-][- int-]
-   type ('a[-, 'b-][-, 'c-]) u =
-     {[- mutable-] a : 'a; b : [-'b-]{+int+};[- c : 'c;-] }
-   type v =
-     | A of [-int * int-]{+{ a : int; b : int; }+}
-     | B of { a : int; b : [-float-]{+string+}; }
-  
-  [1]
-
-Run api-watcher on the two cmi file with plain word-level diffing option enabled
-
-  $ api-diff --word-diff=plain ref.cmi cur.cmi
-  diff module Cur:
-   type [-'a-] p = [-'a * 'a-]{+int * int+}
-   type t[- =-][- int-]
-   type ('a[-, 'b-][-, 'c-]) u =
-     {[- mutable-] a : 'a; b : [-'b-]{+int+};[- c : 'c;-] }
-   type v =
-     | A of [-int * int-]{+{ a : int; b : int; }+}
-     | B of { a : int; b : [-float-]{+string+}; }
-  
-  [1]
-
-Run api-watcher on the two cmi file with color word-level diffing option enabled
-
-  $ api-diff --word-diff=color ref.cmi cur.cmi
+  $ api-diff --word-diff ref.cmi cur.cmi
   diff module Cur:
    type 'a p = 'a * 'aint * int
    type t = int
@@ -63,5 +35,19 @@ Run api-watcher on the two cmi file with color word-level diffing option enabled
    type v =
      | A of int * int{ a : int; b : int; }
      | B of { a : int; b : floatstring; }
+  
+  [1]
+
+Run api-watcher on the two cmi file with plain and word-level diffing flags enabled
+
+  $ api-diff --word-diff --plain ref.cmi cur.cmi
+  diff module Cur:
+   type [-'a-] p = [-'a * 'a-]{+int * int+}
+   type t[- =-][- int-]
+   type ('a[-, 'b-][-, 'c-]) u =
+     {[- mutable-] a : 'a; b : [-'b-]{+int+};[- c : 'c;-] }
+   type v =
+     | A of [-int * int-]{+{ a : int; b : int; }+}
+     | B of { a : int; b : [-float-]{+string+}; }
   
   [1]